+++ src/stlink.ts
@@ -5,15 +5,22 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { EventEmitter } from 'events';
 
-// Path of the top-level STM32CubeIDE installation directory, os-dependant
-const ST_DIR = (
-    os.platform() === 'win32' ? 'c:\\ST' :
-        os.platform() === 'darwin' ? '/Applications/STM32CubeIDE.app/Contents/Eclipse' : '/opt/st'
-    );
+function get_ST_DIR() {
+    switch (os.platform()) {
+        case 'win32':
+            return 'C:\\ST';
+        case 'darwin':
+            return '/Applications/STM32CubeIDE.app/Contents/Eclipse';
+        default:
+            const dirName = (process.env.HOME || os.homedir()) + '/st';
+            return fs.existsSync(dirName) ? dirName : '/opt/st';
+    }
+}
 
-const SERVER_EXECUTABLE_NAME = (
-    os.platform() === 'win32' ? 'ST-LINK_gdbserver.exe' : 'ST-LINK_gdbserver'
-);
+// Path of the top-level STM32CubeIDE installation directory, os-dependant
+const ST_DIR = get_ST_DIR();
+const SERVER_EXECUTABLE_NAME = os.platform() === 'win32' ? 'ST-LINK_gdbserver.exe' : 'ST-LINK_gdbserver';
+const LD_PATH_NAME = os.platform() === 'darwin' ? 'DYLD_FALLBACK_LIBRARY_PATH' : 'LD_LIBRARY_PATH';
 
 const STMCUBEIDE_REGEX = /^STM32CubeIDE_(.+)$/;
 // Example: c:\ST\STM32CubeIDE_1.5.0\
@@ -58,6 +65,7 @@ function resolveCubePath(dirSegments, regex, suffix, executable = '')
 export class STLinkServerController extends EventEmitter implements GDBServerController {
     public readonly name: string = 'ST-LINK';
     public readonly portsNeeded: string[] = ['gdbPort'];
+    private saveLdPath: string | undefined = undefined;
 
     private args: ConfigurationArguments;
     private ports: { [name: string]: number };
@@ -100,9 +108,9 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
 
     public launchCommands(): string[] {
         const commands = [
-            'interpreter-exec console "monitor reset halt"',
+            'interpreter-exec console "monitor reset"',
             'target-download',
-            'interpreter-exec console "monitor reset halt"',
+            'interpreter-exec console "monitor reset"',
             'enable-pretty-printing'
         ];
         return commands;
@@ -110,7 +118,7 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
 
     public attachCommands(): string[] {
         const commands = [
-            'interpreter-exec console "monitor halt"',
+            'interpreter-exec console "monitor"',
             'enable-pretty-printing'
         ];
         return commands;
@@ -118,7 +126,7 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
 
     public restartCommands(): string[] {
         const commands: string[] = [
-            'interpreter-exec console "monitor reset halt"'
+            'interpreter-exec console "monitor reset"'
         ];
         return commands;
     }
@@ -126,9 +134,13 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
     public swoAndRTTCommands(): string[] {
         return [];
     }
+
     public serverExecutable(): string {
-        if (this.args.serverpath) { return this.args.serverpath; }
-        else { return resolveCubePath([STLinkServerController.getSTMCubeIdeDir(), 'plugins'], GDB_REGEX, 'tools/bin', SERVER_EXECUTABLE_NAME); }
+        if (this.args.serverpath) {
+            return this.args.serverpath;
+        } else {
+            return resolveCubePath([STLinkServerController.getSTMCubeIdeDir(), 'plugins'], GDB_REGEX, 'tools/bin', SERVER_EXECUTABLE_NAME);
+        }
     }
 
     public serverArguments(): string[] {
@@ -136,11 +148,12 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
 
         let serverargs = ['-p', gdbport.toString()];
 
-        // The -cp parameter is mandatory and either STM32CubeProgrammer or STM32CubeProgrammer must be installed
-        if (this.args.stm32cubeprogrammer) {
-            serverargs.push('-cp', this.args.stm32cubeprogrammer);
+        // The -cp parameter is mandatory and either STM32CubeIDE or STM32CubeProgrammer must be installed
+        let stm32cubeprogrammer = this.args.stm32cubeprogrammer;
+        if (stm32cubeprogrammer) {
+            serverargs.push('-cp', stm32cubeprogrammer);
         } else {
-            let stm32cubeprogrammer = resolveCubePath([STLinkServerController.getSTMCubeIdeDir(), 'plugins'], PROG_REGEX, 'tools/bin');
+            stm32cubeprogrammer = resolveCubePath([STLinkServerController.getSTMCubeIdeDir(), 'plugins'], PROG_REGEX, 'tools/bin');
             // Fallback to standalone programmer if no STMCube32IDE is installed:
             if (!stm32cubeprogrammer) {
                 if (os.platform() === 'win32') {
@@ -154,6 +167,8 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
             serverargs.push('-cp', stm32cubeprogrammer);
         }
 
+        this.setLDPath(stm32cubeprogrammer);
+         
         if (this.args.interface !== 'jtag') {
             serverargs.push('--swd');
         }
@@ -164,9 +179,29 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
         
         if (this.args.serverArgs) {
             serverargs = serverargs.concat(this.args.serverArgs);
+        } else {
+            serverargs.push('--halt');      // User may want to not use this. controlled via serverArgs
         }
 
         return serverargs;
+    }
+
+    private setLDPath(programmer: string) {
+        if (os.platform() !== 'win32') {
+            const exe = path.dirname(this.serverExecutable());
+            const paths = [];
+            for (const dir of [programmer, exe]) {
+                for (const subdir of ['', '../lib']) {
+                    const p = subdir ? path.join(dir, subdir) : dir;
+                    if (fs.existsSync(p)) {
+                        paths.push(p);
+                    }
+                }
+            }
+            const ldPath = paths.join(':');
+            this.saveLdPath = process.env[LD_PATH_NAME];
+            process.env[LD_PATH_NAME] = this.saveLdPath ? (this.saveLdPath + ':' + ldPath) : ldPath;
+        }
     }
 
     public initMatch(): RegExp {
@@ -175,6 +210,13 @@ export class STLinkServerController extends EventEmitter implements GDBServerCon
 
     public serverLaunchStarted(): void {}
     public serverLaunchCompleted(): void {
+        if (os.platform() !== 'win32') {
+            if (this.saveLdPath === undefined) {
+                delete process.env[LD_PATH_NAME];
+            } else {
+                process.env[LD_PATH_NAME] = this.saveLdPath;
+            }
+        }
     }
     
     public debuggerLaunchStarted(): void {}