+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ChangeLog
+#V1.1.1
+* For STLink Gdb-server and Cube-Programmer, set `LD_LIBRARY_PATH` on Linux and equivalent `DYLD_FALLBACK_LIBRARY_PATH` on MacOS
+* Experimental (Unstable): Chained configurations for multi-core/multi-processor/multi-session debugging
+
 #V1.1.0
 * Version numbering change. Now that VSCode extensions are allowed to publish pre-releases via the marketplace, we will be moving to a new version numbering schem. A version number is (following SemVer standard) major.minor.patch. The general SemVer standard also allows a suffix for alpha, beta, etc., but this is not supported by the marketplace. Instead, they recommend that ODD minor versions be pre-releases and EVEN ones for release versions. This versioning scheme may change in the future. See: https://code.visualstudio.com/api/working-with-extensions/publishing-extension#prerelease-extensions
   This will be the first pre-release coming to you direct from the marketplace.